+++ water/lib/Dialect/Wave/Transforms/Utils.cpp
@@ -21,15 +21,6 @@
 
 using namespace mlir;
 
-wave::WaveHyperparameterAttr wave::getHyperparameters(Operation *op) {
-  for (Operation *current = op; current; current = current->getParentOp()) {
-    if (auto hyperparams = current->getAttrOfType<WaveHyperparameterAttr>(
-            WaveDialect::kHyperparameterAttrName))
-      return hyperparams;
-  }
-  return nullptr;
-}
-
 llvm::LogicalResult wave::collectWaveConstraints(
     Operation *top, llvm::DenseMap<Operation *, Attribute> &constraints) {
   auto *waveDialect = top->getContext()->getLoadedDialect<wave::WaveDialect>();
